add character overview endpoint
Add get character overview logic to HighScoreService. First gets the
characterId based on the characterName. If it does not exist, return not
found. Add method to repository to return xp drops for a given
characterId that exists after a cutoff data, but if no row exists for a
given skill during that timeframe, return the single most recent row for
that skillId. This allows us to always have data to display the up to
date xp for all skills on the character page.